PGB-9, PGB-10

+++ ui.py
@@ -1,4 +1,4 @@
-from PyQt5 import uic, QtWidgets, QtGui
+from PyQt5 import uic, QtWidgets
 from NewTask import NewTaskDialog
 import data_db
 
@@ -13,35 +13,53 @@ class MainWindow(QtWidgets.QMainWindow):
         self.CreateTaskButton.clicked.connect(self.create_task)
         self.DeleteTaskButton.clicked.connect(self.delete_task)
         self.HideDeactivCheckBox.clicked.connect(self.refresh_list)
-
-    def showEvent(self, event):
-        pass
+        self.ActiveTaskButton.clicked.connect(self.activate_task)
+        self.DeactivTaskButton.clicked.connect(self.deactivate_task)
+        self.TaskTable.doubleClicked.connect(self.edit_task)
 
     def delete_task(self):
-        row = self.TaskTable.currentRow()
-        if row == -1:
-            msg = QtWidgets.QMessageBox()
-            msg.setIcon(QtWidgets.QMessageBox.Information)
-            msg.setText('Выберите в списке задачу для удаления')
-            msg.setWindowTitle('Ошибка')
-            msg.exec_()
-        else:
-            msg = QtWidgets.QMessageBox()
-            msg.setIcon(QtWidgets.QMessageBox.Question)
-            msg.setText('Вы действительно хотите удалить эту задачу?')
-            msg.setWindowTitle('Подтверждение')
-            msg.setDefaultButton(QtWidgets.QMessageBox.Yes)
-            msg.setEscapeButton(QtWidgets.QMessageBox.No)
+        if self.check_selected_task():
+            row = self.TaskTable.currentRow()
+            if row == -1:
+                msg = QtWidgets.QMessageBox()
+                msg.setIcon(QtWidgets.QMessageBox.Information)
+                msg.setText('Выберите в списке задачу для удаления')
+                msg.setWindowTitle('Ошибка')
+                msg.exec_()
+            else:
+                msg = QtWidgets.QMessageBox()
+                msg.setIcon(QtWidgets.QMessageBox.Question)
+                msg.setText('Вы действительно хотите удалить эту задачу?')
+                msg.setWindowTitle('Подтверждение')
+                msg.setDefaultButton(QtWidgets.QMessageBox.Yes)
+                msg.setEscapeButton(QtWidgets.QMessageBox.No)
 
-            msg.setStandardButtons(QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.No)
-            retval = msg.exec_()
-            if retval == QtWidgets.QMessageBox.Yes:
-                self.db.delete_task(self.TaskTable.item(row, 0).text())
-                self.refresh_list()
+                msg.setStandardButtons(QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.No)
+                retval = msg.exec_()
+                if retval == QtWidgets.QMessageBox.Yes:
+                    self.db.delete_task(self.TaskTable.item(row, 0).text())
+                    self.refresh_list()
 
     def create_task(self):
         dialog = NewTaskDialog(self)
         dialog.exec_()
+
+    def edit_task(self):
+        pass
+
+    def activate_task(self):
+        if self.check_selected_task():
+            row = self.TaskTable.currentRow()
+            id_task = self.TaskTable.item(row, 0).text()
+            self.db.activate_task(id_task)
+            self.refresh_list()
+
+    def deactivate_task(self):
+        if self.check_selected_task():
+            row = self.TaskTable.currentRow()
+            id_task = self.TaskTable.item(row, 0).text()
+            self.db.deactivate_task(id_task)
+            self.refresh_list()
 
     def closeEvent(self, event):
         pass
@@ -76,3 +94,15 @@ class MainWindow(QtWidgets.QMainWindow):
 
         self.TaskTable.resizeColumnsToContents()
         self.TaskTable.resizeRowsToContents()
+
+    def check_selected_task(self):
+        row = self.TaskTable.currentRow()
+        if row == -1:
+            msg = QtWidgets.QMessageBox()
+            msg.setIcon(QtWidgets.QMessageBox.Information)
+            msg.setText("Выберите задачу из списка")
+            msg.setWindowTitle("Ошибка")
+            msg.exec_()
+            return False
+        else:
+            return True
